extracted domain logic from prototype and from transaction
